Make fuel-indexer-api-server standalone
