+++ examples/simple-wasm/simple-wasm/Cargo.toml
@@ -8,9 +8,9 @@ publish = false
 crate-type = ['cdylib']
 
 [dependencies]
-fuel-indexer-macros = { version = "0.1", path = "../../../fuel-indexer-macros", default-features = false }
-fuel-indexer-plugin = { version = "0.1", path = "../../../fuel-indexer-plugin" }
-fuel-indexer-schema = { version = "0.1", path = "../../../fuel-indexer-schema", default-features = false }
+fuel-indexer-macros = { path = "../../../fuel-indexer-macros", default-features = false }
+fuel-indexer-plugin = { path = "../../../fuel-indexer-plugin" }
+fuel-indexer-schema = { path = "../../../fuel-indexer-schema", default-features = false }
 fuel-tx = "0.18"
 fuels-core = "0.26"
 fuels-types = "0.26"
