+++ fuel-indexer-api-server/Cargo.toml
@@ -5,23 +5,17 @@ edition = "2021"
 license = "BUSL-1.1"
 publish = false
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[lib]
-name = "api_server"
-path = "src/lib.rs"
-
-[[bin]]
-name = "fuel-indexer-api-server"
-path = "src/main.rs"
-
 [dependencies]
 anyhow = "1.0"
 async-std = "1"
-axum = "0.4"
-fuel-indexer = { version = "0.1", path = "../fuel-indexer" }
-fuel-indexer-lib = { version = "0.1", path = "../fuel-indexer-lib" }
-fuel-indexer-schema = { version = "0.1", path = "../fuel-indexer-schema", features = ["db-models"] }
+axum = { version = "0.4", features = ["multipart"] }
+fuel-indexer-database = { path = "../fuel-indexer-database" }
+fuel-indexer-database-types = { path = "../fuel-indexer-database/database-types" }
+fuel-indexer-lib = { path = "../fuel-indexer-lib" }
+fuel-indexer-schema = { path = "../fuel-indexer-schema", features = ["db-models"] }
+http = "0.2"
+hyper = { version = "0.14", features = ["client", "http2", "http1", "runtime" ]}
+hyper-tls = "0.5"
 serde = { version = "1.0", features = ["derive"] }
 serde_json = { version = "1.0", features = ["raw_value"] }
 serde_yaml = "0.8"
