+++ fuel-indexer-api-server/src/bin/main.rs
@@ -1,8 +1,6 @@
 use anyhow::Result;
-use fuel_indexer::{
-    config::{IndexerArgs, Parser},
-    GraphQlApi, IndexerConfig,
-};
+use fuel_indexer_api_server::api::GraphQlApi;
+use fuel_indexer_lib::config::{IndexerArgs, IndexerConfig, Parser};
 use tracing::info;
 use tracing_subscriber::filter::EnvFilter;
 
@@ -23,7 +21,7 @@ pub async fn main() -> Result<()> {
     let opts = IndexerArgs::from_args();
 
     let config = match &opts.config {
-        Some(path) => IndexerConfig::from_file(path).await?,
+        Some(path) => IndexerConfig::from_file(path)?,
         None => IndexerConfig::from_opts(opts.clone()),
     };
