+++ fuel-indexer-api-server/src/lib.rs
@@ -1,1 +1,2 @@
-
+pub mod api;
+mod uses;
