+++ fuel-indexer-database/Cargo.toml
@@ -7,7 +7,7 @@ edition = "2021"
 
 [dependencies]
 fuel-indexer-database-types = { path = "./database-types" }
-fuel-indexer-lib = { version = "0.1", path = "../fuel-indexer-lib" }
+fuel-indexer-lib = { path = "../fuel-indexer-lib" }
 fuel-indexer-postgres = { path = "./postgres" }
 fuel-indexer-sqlite = { path = "./sqlite" }
 sqlx = { version = "0.6" }
