+++ fuel-indexer-database/postgres/Cargo.toml
@@ -7,6 +7,6 @@ edition = "2021"
 
 [dependencies]
 fuel-indexer-database-types = { path = "../database-types" }
-fuel-indexer-lib = { version = "0.1", path = "../../fuel-indexer-lib" }
+fuel-indexer-lib = { path = "../../fuel-indexer-lib" }
 sqlx = { version = "0.6", features = ["runtime-tokio-rustls", "postgres", "offline"] }
 tracing = "0.1"
