+++ fuel-indexer-lib/Cargo.toml
@@ -8,8 +8,12 @@ edition = "2021"
 
 [dependencies]
 anyhow = { version = "1.0", default-features = false }
+# async-std = "1"
+# async-trait = "0.1"
+clap = { version = "3.1", features = ["cargo", "derive", "env"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_yaml = "0.8"
 sha2 = "0.9"
+strum = { version = "0.24", default-features = false, features = ["derive"] }
 tokio = { version = "1.8", features = ["time", "rt"] }
 tracing = "0.1"
