+++ fuel-indexer-lib/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod config;
 pub mod defaults;
 pub mod manifest;
 pub mod utils;
