+++ fuel-indexer-macros/Cargo.toml
@@ -13,8 +13,8 @@ development = ["fuels", "fuel-indexer-plugin"]
 proc-macro = true
 
 [dependencies]
-fuel-indexer-lib = { version = "0.1", path = "../fuel-indexer-lib", default-features = false }
-fuel-indexer-schema = { version = "0.1", path = "../fuel-indexer-schema", default-features = false }
+fuel-indexer-lib = { path = "../fuel-indexer-lib", default-features = false }
+fuel-indexer-schema = { path = "../fuel-indexer-schema", default-features = false }
 fuel-tx = "0.18"
 fuels-core = { version = "0.26" }
 fuels-types = { version = "0.26" }
@@ -29,7 +29,7 @@ sha2 = "0.9.5"
 syn = { version = "1.0", features = ["full"] }
 
 [dev-dependencies]
-fuel-indexer-plugin = { version = "0.1", path = "../fuel-indexer-plugin" }
+fuel-indexer-plugin = { path = "../fuel-indexer-plugin" }
 fuels = { version = "0.26" }
 trybuild = "1.0"
