+++ fuel-indexer-plugin/Cargo.toml
@@ -10,5 +10,5 @@ publish = false
 crate-type = ['rlib']
 
 [dependencies]
-fuel-indexer-lib = { version = "0.1", path = "../fuel-indexer-lib", default-features = false }
-fuel-indexer-schema = { version = "0.1", path = "../fuel-indexer-schema", default-features = false }
+fuel-indexer-lib = { path = "../fuel-indexer-lib", default-features = false }
+fuel-indexer-schema = { path = "../fuel-indexer-schema", default-features = false }
