+++ fuel-indexer-tests/components/web/Cargo.toml
@@ -18,7 +18,7 @@ actix-web = { version = "4", default-features = false }
 anyhow = { version = "1.0", default-features = false }
 async-std = "1"
 clap = { version = "3.1", features = ["cargo", "derive", "env"] }
-fuel-indexer-lib = { version = "0.1", path = "../../../fuel-indexer-lib" }
+fuel-indexer-lib = { path = "../../../fuel-indexer-lib" }
 fuel-indexer-tests = { path = "./../../../fuel-indexer-tests" }
 fuel-tx = "0.18"
 fuel-types = "0.5"
