+++ fuel-indexer/Cargo.toml
@@ -15,11 +15,12 @@ bincode = "1.3.3"
 cfg-if = "1.0"
 clap = { version = "3.1", features = ["cargo", "derive", "env"] }
 fuel-gql-client = { version = "0.10" }
-fuel-indexer-database = { version = "0.1", path = "../fuel-indexer-database" }
+fuel-indexer-api-server = { path = "../fuel-indexer-api-server", optional = true }
+fuel-indexer-database = { path = "../fuel-indexer-database" }
 fuel-indexer-database-types = { path = "../fuel-indexer-database/database-types" }
-fuel-indexer-lib = { version = "0.1", path = "../fuel-indexer-lib" }
+fuel-indexer-lib = { path = "../fuel-indexer-lib" }
 fuel-indexer-postgres = { path = "../fuel-indexer-database/postgres" }
-fuel-indexer-schema = { version = "0.1", path = "../fuel-indexer-schema", features = ["db-models"] }
+fuel-indexer-schema = { path = "../fuel-indexer-schema", features = ["db-models"] }
 fuel-indexer-sqlite = { path = "../fuel-indexer-database/sqlite" }
 fuel-tx = "0.18"
 fuel-types = "0.5"
@@ -49,4 +50,5 @@ itertools = { version = "0.10" }
 rand = "0.8"
 
 [features]
-default = []
+default = ["api-server"]
+api-server = ["fuel-indexer-api-server"]
