--- fuel-indexer/src/test_data/contracts-abi.json
@@ -1,88 +1,0 @@
-{
-  "types": [
-    {
-      "typeId": 0,
-      "type": "b256",
-      "components": null,
-      "typeParameters": null
-    },
-    {
-      "typeId": 1,
-      "type": "struct AnotherEvent",
-      "components": [
-        {
-          "name": "id",
-          "type": 3,
-          "typeArguments": null
-        },
-        {
-          "name": "account",
-          "type": 0,
-          "typeArguments": null
-        },
-        {
-          "name": "hash",
-          "type": 0,
-          "typeArguments": null
-        }
-      ],
-      "typeParameters": null
-    },
-    {
-      "typeId": 2,
-      "type": "struct SomeEvent",
-      "components": [
-        {
-          "name": "id",
-          "type": 3,
-          "typeArguments": null
-        },
-        {
-          "name": "account",
-          "type": 0,
-          "typeArguments": null
-        }
-      ],
-      "typeParameters": null
-    },
-    {
-      "typeId": 3,
-      "type": "u64",
-      "components": null,
-      "typeParameters": null
-    }
-  ],
-  "functions": [
-    {
-      "inputs": [
-        {
-          "name": "num",
-          "type": 3,
-          "typeArguments": null
-        }
-      ],
-      "name": "gimme_someevent",
-      "output": {
-        "name": "",
-        "type": 2,
-        "typeArguments": null
-      }
-    },
-    {
-      "inputs": [
-        {
-          "name": "num",
-          "type": 3,
-          "typeArguments": null
-        }
-      ],
-      "name": "gimme_anotherevent",
-      "output": {
-        "name": "",
-        "type": 1,
-        "typeArguments": null
-      }
-    }
-  ],
-  "loggedTypes": []
-}
